-ažuriraj servisne klase da zadovoljavaju testove

+++ angular/frontend/src/app/nutrition/nutrition-service/nutrition.service.ts
@@ -20,6 +20,9 @@ export class NutritionService {
   }
 
   deleteFood(foodID: number) {
+    if (isNaN(foodID)) {
+      throw new Error('Invalid food ID!');
+    }
     return this.httpClient.delete<HttpResponse<object>>(
       `${this.endpoint}${foodID}`,
       {
@@ -29,6 +32,11 @@ export class NutritionService {
   }
 
   addToNutrition(body: IAddFoodToNutrition) {
+    if (body.id !== undefined && body.id.length == 0) {
+      throw new Error('ID is empty!');
+    } else if (body.name !== undefined && body.name.length == 0) {
+      throw new Error('Name is empty!');
+    }
     return this.httpClient.post<HttpResponse<object>>(
       `${this.endpoint}`,
       body,
@@ -41,6 +49,11 @@ export class NutritionService {
   updateQuantity(body: IUpdateFoodsBody) {
     if (!body.foods.length) {
       throw new Error('No foods to update!');
+    } else if (body.foods.some((food) => food.quantity <= 0)) {
+      throw new Error('Quantity of 0 or less is not allowed!');
+    }
+    if (body.foods.some((food) => isNaN(food.id))) {
+      throw new Error('One or more foods have invalid ID!');
     }
     return this.httpClient.put<HttpResponse<object>>(`${this.endpoint}`, body, {
       observe: 'response',
