+++ angular/frontend/src/app/workout-plan/workout-plan-service/workout-plan.service.ts
@@ -14,6 +14,9 @@ import { IWorkoutPlanDetails } from 'src/interfaces/iworkout-plan-details';
 })
 export class WorkoutPlanService {
   private endpoint: string = `${environment.apiUrl}workout-plans`;
+
+  constructor(private httpClient: HttpClient) {}
+
   getAllWorkoutPlans(): Observable<IWorkoutPlan[]> {
     return this.httpClient.get<IWorkoutPlan[]>(this.endpoint, {
       observe: 'body',
@@ -22,6 +25,9 @@ export class WorkoutPlanService {
   }
 
   getDetails(id: number) {
+    if (isNaN(id)) {
+      throw new Error('ID is invalid!');
+    }
     return this.httpClient.get<IWorkoutPlanDetails>(`${this.endpoint}/${id}`, {
       observe: 'body',
       responseType: 'json',
@@ -29,12 +35,22 @@ export class WorkoutPlanService {
   }
 
   deleteWorkoutPlan(deleteWorkoutPlanData: IDeleteWorkoutPlan) {
+    if (isNaN(deleteWorkoutPlanData.id)) {
+      throw new Error('ID is invalid!');
+    }
     return this.httpClient.delete<HttpResponse<object>>(`${this.endpoint}`, {
       body: deleteWorkoutPlanData,
       observe: 'response',
     });
   }
+
   deleteExercise(workoutPlanId: number, deleteExercise: IDeleteExercise) {
+    if (isNaN(workoutPlanId)) {
+      throw new Error('Invalid workout plan ID!');
+    }
+    if (!deleteExercise.name.length) {
+      throw new Error('Exercise name not valid!');
+    }
     return this.httpClient.delete<HttpResponse<object>>(
       `${this.endpoint}/${workoutPlanId}`,
       {
@@ -46,6 +62,9 @@ export class WorkoutPlanService {
   createWorkoutPlan(
     workoutPlanData: ICreateWorkoutPlan
   ): Observable<HttpResponse<object>> {
+    if (!workoutPlanData.title.length) {
+      throw new Error('Title is empty!');
+    }
     return this.httpClient.post(this.endpoint, workoutPlanData, {
       observe: 'response',
       responseType: 'json',
@@ -53,6 +72,12 @@ export class WorkoutPlanService {
   }
 
   addExercise(workoutPlanId: number, exercise: IAddExercise) {
+    if (isNaN(workoutPlanId)) {
+      throw new Error('Invalid workout plan ID!');
+    }
+    if (!exercise.name.length) {
+      throw new Error('Exercise name not valid!');
+    }
     return this.httpClient.post<HttpResponse<object>>(
       `${this.endpoint}/${workoutPlanId}`,
       exercise,
@@ -61,6 +86,4 @@ export class WorkoutPlanService {
       }
     );
   }
-
-  constructor(private httpClient: HttpClient) {}
 }
